requirements.txt
# torch 
# torchvision 
opencv-python 
jupyter notebook
prettytable
tqdm
torchsummary

